IK solver convergence updates

+++ python/rational_linkages/MotionFactorization.py
@@ -283,7 +283,8 @@ class MotionFactorization(RationalCurve):
         if t_param_joint0 == 0.0:
             t_param_joint0 = np.finfo(float).eps
 
-        angle = 2 * np.arctan(np.sqrt(self.dq_axes[0].p.norm()) / t_param_joint0)
+        angle = 2 * np.arctan(np.float64(
+            np.sqrt(self.dq_axes[0].p.norm()) / t_param_joint0))
 
         # normalize angle to [0, 2*pi]
         if angle < 0:
